% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{gad7}
\alias{gad7}
\alias{gad7.rev}
\title{Convert GAD-7 scores to an anxiety common metric}
\usage{
gad7(x)
gad7.rev(x)
}
\arguments{
\item{x}{A scalar or vector of scores that should be converted.}
}
\description{
This function allows you to convert scores of the 7-item Generalized Anxiety Disorder (GAD-7) questionnaire
to an anxiety "common metric" score. Common metric scores
can be "reconverted" to the original scale by appending \code{.rev} to the respective
converter function.
}
\details{
Conversions are conducted based on the IRT model developed in
Fischer et al. (\href{https://doi.org/10.1007/s11136-013-0599-y}{2013}; supplement).
Common metric scores \emph{θ} are normed so that values of 50 indicate the general
population mean, with a population standard deviation of 10.
}
\examples{
\dontrun{
x <- runif(100,0,30)
gad7(x)
}
}
