% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{hadsa}
\alias{hadsa}
\alias{hadsa.rev}
\title{Convert HADS-A scores to an anxiety common metric}
\usage{
hadsa(x)
hadsa.rev(x)
}
\arguments{
\item{x}{A scalar or vector of scores that should be converted.}
}
\description{
This function allows you to convert scores of the Hospital Anxiety and Depression Scale anxiety subscale
to an anxiety "common metric" score. Common metric scores
can be "reconverted" to the original scale by appending \code{.rev} to the respective
converter function.
}
\details{
Conversions are conducted based on the IRT model developed in
Fischer et al. (\href{https://doi.org/10.1007/s11136-013-0599-y}{2013}; supplement).
Common metric scores \emph{θ} are normed so that values of 50 indicate the general
population mean, with a population standard deviation of 10.
}
\examples{
\dontrun{
x <- runif(100,0,30)
hadsa(x)
}
}
