% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{hamd}
\alias{hamd}
\alias{hamd.rev}
\title{Convert HAM-D/HDRS-17 scores to a depression common metric}
\usage{
hamd(x)
hamd.rev(x)
}
\arguments{
\item{x}{A scalar or vector of scores that should be converted.}
}
\description{
This function allows you to convert scores of the 17-item Hamilton Depression Rating Scale
to a depression "common metric" score. Common metric scores
can be "reconverted" to the original scale by appending \code{.rev} to the respective
converter function.
}
\details{
Conversions are conducted using the crosswalk tables provided in
Wahl et al. (\href{https://doi.org/10.1016/j.jclinepi.2013.04.019}{2014}; supplement).
Common metric scores \emph{θ} are normed so that values of 50 indicate the general
population mean, with a population standard deviation of 10.

The HAM-D/HDRS-17 is not part of the original depression common metric model.
To convert scores, values are first transformed to BDI-I equivalents using the
equipercentile linking crosswalk developed by Furukawa et al.
(\href{https://doi.org/10.1017/S2045796019000088}{2020}). Transformed scores should therefore be
interpreted cautiously.
}
\examples{
\dontrun{
x <- runif(100,0,30)
hamd(x)
}
}
