% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{madrs}
\alias{madrs}
\alias{madrs.rev}
\title{Convert MÅDRS scores to a depression common metric}
\usage{
madrs(x)
madrs.rev(x)
}
\arguments{
\item{x}{A scalar or vector of scores that should be converted.}
}
\description{
This function allows you to convert scores of the Montgomery-Åsberg Depression Rating Scale
to a depression "common metric" score. Common metric scores
can be "reconverted" to the original scale by appending \code{.rev} to the respective
converter function.
}
\details{
Conversions are conducted using the crosswalk tables provided in
Wahl et al. (\href{https://doi.org/10.1016/j.jclinepi.2013.04.019}{2014}; supplement).
Common metric scores \emph{θ} are normed so that values of 50 indicate the general
population mean, with a population standard deviation of 10.

The MÅDRS is not part of the original depression common metric model.
To convert scores, values are transformed to HAM-D and then BDI-I equivalents using the
equipercentile linking crosswalk developed by Furukawa et al.
(\href{https://doi.org/10.1017/S2045796019000088}{2020}), and the IRT-based crosswalk by Carmody et al.
(\href{10.1016/j.euroneuro.2006.04.008}{2006}). Transformed scores should therefore be
interpreted cautiously.
}
\examples{
\dontrun{
x <- runif(100,0,30)
madrs(x)
}
}
