% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{qids}
\alias{qids}
\alias{qids.rev}
\title{Convert QIDS-SR scores to a depression common metric}
\usage{
qids(x)
qids.rev(x)
}
\arguments{
\item{x}{A scalar or vector of scores that should be converted.}
}
\description{
This function allows you to convert scores of the 16-item Quick Inventory of Depressive Symptomatology (QIDS-SR)
to a depression "common metric" score. Common metric scores
can be "reconverted" to the original scale by appending \code{.rev} to the respective
converter function.
}
\details{
Conversions are conducted using the crosswalk tables provided in
Wahl et al. (\href{https://doi.org/10.1016/j.jclinepi.2013.04.019}{2014}; supplement).
Common metric scores \emph{θ} are normed so that values of 50 indicate the general
population mean, with a population standard deviation of 10.

The QIDS-SR is not part of the original depression common metric model.
To convert scores, values are transformed to HAM-D and then BDI-I equivalents using the
equipercentile linking crosswalk developed by Furukawa et al.
(\href{https://doi.org/10.1017/S2045796019000088}{2020}), and the IRT-based crosswalk by Rush et al.
(\href{https://doi.org/10.1016/S0006-3223(02)01866-8}{2003}). Transformed scores should therefore be
interpreted cautiously.
}
\examples{
\dontrun{
x <- runif(100,0,30)
qids(x)
}
}
